Set up folder structer, added some styles and first option
