+++ assets/modal-form.scss
@@ -1,0 +1,28 @@
+.modal-form {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  height: 300px;
+
+  .radio-label {
+    margin: 10px 10px;
+    padding: 15px;
+    border: 3px solid $dark-blue;
+    color: $dark-blue;
+  }
+
+  input[type="radio"] {
+    display: none;
+  }
+
+  input[type="radio"]:checked + label {
+    background-color: $dark-blue;
+    color: $white;
+  }
+
+  .modal-form__group {
+    display: flex;
+    justify-content: center;
+  }
+}
