+++ assets/modal.scss
@@ -1,0 +1,34 @@
+.modalBox {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  width: 500px;
+  max-width: 100%;
+  background-color: $white;
+  box-shadow: 4px 4px 4px $shadow-blue;
+  border: 2px solid $dark-blue;
+
+  .modalBox__header {
+    width: 100%;
+    padding: 10px 10px;
+    background-color: $dark-blue;
+    display: flex;
+    justify-content: space-between;
+  }
+
+  .modalBox__header-text {
+    color: $white;
+    padding: 0;
+    margin: 0;
+  }
+
+  .modalBox__close {
+    color: $white;
+  }
+
+  .modalBox__content {
+    padding: 10px 20px;
+    color: $dark-blue;
+  }
+}
