+++ assets/variables.scss
@@ -1,0 +1,14 @@
+// VARIABLES
+
+// Colors
+$white: #fff;
+$black: #000;
+
+$shadow-blue: rgba(23, 39, 105, 0.253);
+
+$rose: #eda9ab;
+$peach: #f28a8a;
+$red-peach: #f75454;
+$turquoise: #1bb6af;
+$blue: #088bbe;
+$dark-blue: #172869;
