--- main.css
@@ -1,48 +1,0 @@
-* {
-  box-sizing: border-box;
-}
-header {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-main {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  height: 300px;
-}
-
-.modalBox {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  width: 400px;
-  background-color: rgb(199, 199, 199);
-  box-shadow: 2px 2px 3px rgba(0, 0, 0, 0.2);
-}
-
-.modalBox__header {
-  width: 100%;
-  padding: 5px 10px;
-  background-color: #3d3d3d;
-  display: flex;
-  justify-content: space-between;
-}
-
-.modalBox__header-text {
-  color: #ffffff;
-  padding: 0;
-  margin: 0;
-}
-
-.modalBox__close {
-  color: #ffffff;
-}
-
-.modalBox__content {
-  padding: 5px;
-}
